Move fixture tracking_uri_mock at project-level
